Implement building for big endian
A few lookup tables are created to convert 8 bit integer I/Q data to the
format SoapySDR expects. The lookups index the table with the 16 bit I/Q
combination from the hardware and the big endian part was not
implemented. This patch reorganizes the code a bit to make a clean
little and big endian implementation.