updates for buffer defaults and controls
Use the larger buffer default in librtlsdr itself.
This is the same buffer size we would get with size 0
passed to rtlsdr_read_async() This should mean less
overflows for the default buffer configuration.

Separate buffer counts for the ring buffer and usb.
The size of the ring used for padding does not have
to match the value passed to rtlsdr_read_async()
We can increase the ring size, but libusb can be
sensitive to the increase.